+++ json_wrapper.h
@@ -1,20 +1,21 @@
-//#ifndef JSONWRAPPER_H
-//#define JSONWRAPPER_H
+#ifndef JSONWRAPPER_H
+#define JSONWRAPPER_H
 
 #include <boost/property_tree/json_parser.hpp>
 #include <boost/property_tree/ptree.hpp>
 #include <map>
 #include <iostream>
 
+namespace pt = boost::property_tree;
 using namespace std;
 
-class SFReader () {
+class SF_Reader {
+    map< string, map< string, map< string, map< string, map<string, double> > > > > map_sf;
 
     public:
-    void read_json(string json_file);
+    void read_json(string);
+    double value (string, string, string, string);
+    double error (string, string, string, string);
+};
 
-    private:
-
-    };
-
-//#endif
+#endif
